update ChildReport and ChildReportScores, service, controller

+++ src/main/java/senai/tcc/zupiapi/zupibackend/repositories/ChildReportScoreRepository.java
@@ -3,5 +3,5 @@ package senai.tcc.zupiapi.zupibackend.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import senai.tcc.zupiapi.zupibackend.model.ChildReportScore;
 
-public interface ChildReportScoreRepository extends JpaRepository<ChildReportScore, Integer> {
+public interface ChildReportScoreRepository extends JpaRepository<ChildReportScore, Long> {
 }
